Use directly the FSCK function
